Add files via upload
The-cake-shop
web-lab-project-srvt created by GitHub Classroom
CAKE SHOP WEB PROJECT

This webpage is based on the ordering and selling the cake for each customer. It will display the names and flavours of different cakes along with their prices. Customers may also register onto this webpage. The customer can continue to make payments when they decide to buy a cake of their choice.